fixup

+++ narwhals/_pandas_like/series.py
@@ -484,8 +484,8 @@ class PandasSeriesDateTimeNamespace:
         result = (
             ser.to_numpy().astype("datetime64[D]")
             - (year_start.to_numpy() - 1970).astype("datetime64[Y]")
-        ).astype("int64") + 1
-        dtype = "Int64[pyarrow]" if "pyarrow" in str(ser.dtype) else "int64"
+        ).astype("int32") + 1
+        dtype = "Int64[pyarrow]" if "pyarrow" in str(ser.dtype) else "int32"
         return self._series._from_series(
             self._series._series.__class__(result, dtype=dtype, name=year_start.name)
         )
